+++ bin/darttrack.dart
@@ -3,5 +3,5 @@
 import 'package:darttrack/darttrack.dart';
 
 void main(List<String> arguments) {
-  darttrack(arguments[0], arguments[1]);
+  darttrack(arguments);
 }