+++ bin/demo.dart
@@ -1,0 +1,4 @@
+void main(List<String> args) {
+  print('Demo script is started...');
+  print('Arguments: $args');
+}