+++ lib/src/darttrack.dart
@@ -5,9 +5,9 @@ import 'controller.dart';
 /// Main function of the application
 ///
 /// Start watch [dir] files changes and run [script] when is it happened.
-void darttrack(String dir, String script) {
+void darttrack(List<String> arguments) {
   try {
-    Controller(dir, script);
+    Controller(arguments);
   } on Exception catch (e) {
     print('Controller error: $e');
   }